[#438] Reorganize model classes
The model classes were extracted out of `Onetime` and organized into
submodules and mixins to better encapsulate data logic and improve
readability. This restructures how models are defined and allows
composing behavior through mixins instead of modules. The `RateLimit`,
`Secret`, `Session`, `Customer`, `SplitTest`, and `Subdomain` classes
were moved under this new structure without other changes to their
implementation.
