+++ lib/onetime/models/mixins.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative 'mixins/passphrase'
+require_relative 'mixins/rate_limited'
+require_relative 'mixins/redis_hash'
