+++ lib/onetime/models/mixins/rate_limited.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+
+
+module Onetime::Models
+  module RateLimited
+    def event_incr! event
+      # Uses the external identifier of the implementing class to keep
+      # track of the event count. e.g. sess.external_identifier.
+      OT::RateLimit.incr! external_identifier, event
+    end
+    def event_clear! event
+      OT::RateLimit.clear! external_identifier, event
+    end
+    def external_identifier
+      raise RuntimeError, "TODO: #{self.class}.external_identifier"
+    end
+  end
+end
