+++ lib/onetime/models/modules.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative 'modules/rate_limit'
+require_relative 'modules/feedback'
+require_relative 'modules/entropy'
